feat(PN-12438): update SpecialContacts component
