+++ packages/pn-personafisica-webapp/src/components/Contacts/EmailContactItem.tsx
@@ -27,7 +27,7 @@ import DigitalContactsCard from './DigitalContactsCard';
 import EditDigitalContact from './EditDigitalContact';
 import ExistingContactDialog from './ExistingContactDialog';
 import InsertDigitalContact from './InsertDigitalContact';
-import SpecialContactElem from './SpecialContactElem';
+import SpecialContacts from './SpecialContacts';
 
 interface Props {
   value: string;
@@ -330,7 +330,7 @@ const EmailContactItem: React.FC<Props> = ({
         confirmHandler={deleteConfirmHandler}
         blockDelete={blockDelete}
       />
-      <SpecialContactElem addressType="email" />
+      <SpecialContacts digitalAddresses={digitalAddresses} addressType="email" />
     </DigitalContactsCard>
   );
 };
